Start refactoring for PHP extraction
