+++ service-front/app/src/Common/src/Command/TranslationUpdateCommandFactory.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Common\Command;
 
 use Common\Service\I18n\CatalogueLoader;
+use Common\Service\I18n\Extractors\TwigFactory;
 use Common\Service\I18n\PotGenerator;
-use Common\Service\I18n\TwigCatalogueExtractorFactory;
 use Psr\Container\ContainerInterface;
 
 class TranslationUpdateCommandFactory
@@ -14,7 +14,7 @@ class TranslationUpdateCommandFactory
     public function __invoke(ContainerInterface $container): TranslationUpdateCommand
     {
         return new TranslationUpdateCommand(
-            $container->get(TwigCatalogueExtractorFactory::class),
+            $container->get(TwigFactory::class),
             $container->get(CatalogueLoader::class),
             $container->get(PotGenerator::class),
             [
